add firebase service. Instead of calling the methods in mainActivity, instatiate firebaseservice to staert
